requirements.txt
Django==1.9.1
pinax-theme-bootstrap==7.2.0
django-user-accounts==1.3.1
metron==1.3.7
pinax-eventlog==1.1.1

-e git://github.com/pinax/pinax-lms-activities.git@0518744d78e90cdca36d7b768a5448777b2231aa#egg=pinax-lms-activities
